+++ extensions/core-actions/backup.ts
@@ -1,0 +1,39 @@
+export function filePicker(): Promise<File> {
+  return new Promise((res, rej) => {
+    const input = document.createElement('input');
+    input.type = 'file';
+    input.accept = '.zip';
+
+    input.addEventListener(
+      'change',
+      () => {
+        const file = input.files?.[0];
+        if (file) {
+          res(file);
+        } else {
+          rej(new Error('Unable to pick backup file'));
+        }
+      },
+      { once: true },
+    );
+    input.click();
+  });
+}
+
+export function downloadBlob(blob: Blob, filename: string) {
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = filename;
+  a.addEventListener(
+    'click',
+    () => {
+      setTimeout(() => {
+        URL.revokeObjectURL(url);
+      }, 150);
+    },
+    { once: true },
+  );
+
+  a.click();
+}