+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### HEAD
+
+- **Feature:** Bangle now has new actions that allows the ability to export notes and restore notes.
+
 ### Dec 28, 2021
 
 - **Feature:** Notes now remember your last scroll and selection positions.